requirements.txt
streamlit>=1.36.0
opencv-python-headless>=4.9.0
Pillow>=10.0.0
torch>=2.2.0
torchaudio>=2.2.0
transformers>=4.41.0
sentencepiece>=0.2.0
accelerate>=0.30.0
audiocraft>=1.3.0
pydub>=0.25.1
pytest>=8.2.0
